chore: rename ResultScreen to PollResultScreen
Renamed ResultScreen and its package to PollResultScreen for improved clarity and consistency. Updated all references in ConsensusNavGraph to use the new name.
